refactor: CardSearchService, PersonaService 인터페이스 및 구현체 분리
